requirements.txt
PyQt6
SQLAlchemy
pymysql
python-dotenv
matplotlib
qt-material
pytest
pytest-qt
pytest-cov
